Module ioctl::libc

Modules

consts

funcs

types

Structs

addrinfo

glob_t

ifaddrs

in6_addr

in_addr

ip6_mreq

ip_mreq

pthread_attr_t

rlimit

sockaddr

sockaddr_in

sockaddr_in6

sockaddr_ll

sockaddr_storage

sockaddr_un

stat

timespec

timeval

utimbuf

Enums

DIR

FILE

c_void
Type used to construct void pointers for use with C.

dirent_t

fpos_t

timezone

Constants

AF_INET

AF_INET6

AF_PACKET

AF_UNIX

BUFSIZ

CLOCK_MONOTONIC

CLOCK_REALTIME

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EADV

EAFNOSUPPORT

EAGAIN

EALREADY

EBADE

EBADF

EBADFD

EBADMSG

EBADR

EBADRQC

EBADSLT

EBFONT

EBUSY

ECANCELED

ECHILD

ECHRNG

ECOMM

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EDOTDOT

EDQUOT

EEXIST

EFAULT

EFBIG

EHOSTDOWN

EHOSTUNREACH

EHWPOISON

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

EISNAM

EKEYEXPIRED

EKEYREJECTED

EKEYREVOKED

EL2HLT

EL2NSYNC

EL3HLT

EL3RST

ELIBACC

ELIBBAD

ELIBEXEC

ELIBMAX

ELIBSCN

ELNRNG

ELOOP

EMEDIUMTYPE

EMFILE

EMLINK

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENAVAIL

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOANO

ENOBUFS

ENOCSI

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOKEY

ENOLCK

ENOLINK

ENOMEDIUM

ENOMEM

ENOMSG

ENONET

ENOPKG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTBLK

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTNAM

ENOTRECOVERABLE

ENOTSOCK

ENOTTY

ENOTUNIQ

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERANGE

EREMCHG

EREMOTE

EREMOTEIO

ERESTART

ERFKILL

EROFS

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESRMNT

ESTALE

ESTRPIPE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUCLEAN

EUNATCH

EUSERS

EWOULDBLOCK

EXDEV

EXFULL

EXIT_FAILURE

EXIT_SUCCESS

FILENAME_MAX

FOPEN_MAX

F_DUPFD

F_GETFD

F_GETFL

F_LOCK

F_OK

F_SETFD

F_SETFL

F_TEST

F_TLOCK

F_ULOCK

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

IFF_LOOPBACK

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_RAW

IPPROTO_TCP

IPV6_ADD_MEMBERSHIP

IPV6_DROP_MEMBERSHIP

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_TTL

L_tmpnam

MADV_DOFORK

MADV_DONTFORK

MADV_DONTNEED

MADV_HWPOISON

MADV_MERGEABLE

MADV_NORMAL

MADV_RANDOM

MADV_REMOVE

MADV_SEQUENTIAL

MADV_UNMERGEABLE

MADV_WILLNEED

MAP_32BIT

MAP_ANON

MAP_ANONYMOUS

MAP_DENYWRITE

MAP_EXECUTABLE

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_GROWSDOWN

MAP_LOCKED

MAP_NONBLOCK

MAP_NORESERVE

MAP_POPULATE

MAP_PRIVATE

MAP_SHARED

MAP_STACK

MAP_TYPE

MCL_CURRENT

MCL_FUTURE

MS_ASYNC

MS_INVALIDATE

MS_SYNC

O_ACCMODE

O_APPEND

O_CREAT

O_DSYNC

O_EXCL

O_NOCTTY

O_NONBLOCK

O_RDONLY

O_RDWR

O_RSYNC

O_SYNC

O_TRUNC

O_WRONLY

POSIX_MADV_DONTNEED

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

PROT_EXEC

PROT_GROWSDOWN

PROT_GROWSUP

PROT_NONE

PROT_READ

PROT_WRITE

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_STACK_MIN

RAND_MAX

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_LOCKS

RLIMIT_MEMLOCK

RLIMIT_MSGQUEUE

RLIMIT_NICE

RLIMIT_NLIMITS

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_RSS

RLIMIT_RTPRIO

RLIMIT_RTTIME

RLIMIT_SIGPENDING

RLIMIT_STACK

RLIM_INFINITY

RLIM_SAVED_CUR

RLIM_SAVED_MAX

RUSAGE_CHILDREN

RUSAGE_SELF

RUSAGE_THREAD

R_OK

SEEK_CUR

SEEK_END

SEEK_SET

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGKILL

SIGPIPE

SIGQUIT

SIGSEGV

SIGTERM

SIGTRAP

SIG_IGN

SOCK_DGRAM

SOCK_RAW

SOCK_STREAM

SOL_SOCKET

SO_ACCEPTCONN

SO_BROADCAST

SO_DEBUG

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LINGER

SO_OOBINLINE

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TYPE

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TCP_CONGESTION

TCP_COOKIE_TRANSACTIONS

TCP_CORK

TCP_DEFER_ACCEPT

TCP_FASTOPEN

TCP_INFO

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINTVL

TCP_LINGER2

TCP_MAXSEG

TCP_MD5SIG

TCP_NODELAY

TCP_QUEUE_SEQ

TCP_QUICKACK

TCP_REPAIR

TCP_REPAIR_OPTIONS

TCP_REPAIR_QUEUE

TCP_SYNCNT

TCP_THIN_DUPACK

TCP_THIN_LINEAR_TIMEOUTS

TCP_TIMESTAMP

TCP_USER_TIMEOUT

TCP_WINDOW_CLAMP

TMP_MAX

W_OK

X_OK

_IOFBF

_IOLBF

_IONBF

_PC_NAME_MAX

_SC_2_CHAR_TERM

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_C_VERSION

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_COLL_WEIGHTS_MAX

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_IOV_MAX

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_ONLN

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_REALTIME_SIGNALS

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SIGQUEUE_MAX

_SC_STREAM_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMERS

_SC_TIMER_MAX

_SC_TTY_NAME_MAX

_SC_TZNAME_MAX

_SC_VERSION

_SC_XBS5_ILP32_OFF32

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_LPBIG_OFFBIG

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

Functions

_exit

abs

accept

access

alarm

atexit

atof

atoi

bind

calloc

chdir

chmod

chown

close

closedir

connect

creat

dup

dup2

execv

execve

execvp

exit
Exits the running program in a possibly dangerous manner.

fchmod

fclose

fcntl

fdatasync

fdopen

feof

ferror

fflush

fgetc

fgetpos

fgets

fileno

fopen

fork

fpathconf

fputc

fputs

fread

free

freeifaddrs

freopen

fseek

fsetpos

fstat

fsync

ftell

ftruncate

fwrite

getcwd

getegid

getenv

geteuid

getgid

getgroups

getifaddrs

getlogin

getopt

getpeername

getpgrp

getpid

getppid

getrlimit

getrusage

getsid

getsockname

getuid

if_nametoindex

isalnum

isalpha

isatty

iscntrl

isdigit

isgraph

islower

isprint

ispunct

isspace

isupper

isxdigit

labs

link

listen

lseek

lstat

malloc

memchr

memcmp

mkdir

mkfifo

mlock

mlockall

mmap

mprotect

msync

munlock

munlockall

munmap

nanosleep

open

opendir

pathconf

pause

pclose

perror

pipe

popen

pread

putenv

puts

pwrite

rand

read

readdir_r

readlink

realloc

recv

recvfrom

remove

rename

rewind

rewinddir

rmdir

seekdir

send

sendto

setbuf

setenv

setgid

setpgid

setrlimit

setsid

setsockopt

setuid

setvbuf

shm_open

shm_unlink

shutdown

sleep

socket

srand

stat

strcat

strchr

strcmp

strcoll

strcpy

strcspn

strerror

strlen

strncat

strncmp

strncpy

strpbrk

strrchr

strspn

strstr

strtod

strtok

strtol

strtoul

strxfrm

symlink

sysconf

system

tcgetpgrp

telldir

tmpfile

tolower

toupper

ttyname

ungetc

unlink

unsetenv

usleep

utime

wait

waitpid

wcslen

write

Type Definitions

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

clock_t

dev_t

gid_t

in_addr_t

in_port_t

ino_t

int16_t

int32_t

int64_t

int8_t

intmax_t

intptr_t

mode_t

nlink_t

off_t

pid_t

pthread_t

ptrdiff_t

rlim_t

sa_family_t

sighandler_t

size_t

socklen_t

ssize_t

suseconds_t

time_t

uid_t

uint16_t

uint32_t

uint64_t

uint8_t

uintmax_t

uintptr_t

useconds_t

wchar_t