Crate ioctl

Macros

io
Encode an ioctl command that has no associated data.

ioc
Encode an ioctl command.

ioctl
Declare a wrapper function around an ioctl.

ior
Encode an ioctl command that reads.

iorw
Encode an ioctl command that both reads and writes.

iow
Encode an ioctl command that writes.

Structs

Union_Unnamed16

ff_condition_effect

ff_constant_effect

ff_effect

ff_envelope

ff_periodic_effect

ff_ramp_effect

ff_replay

ff_rumble_effect

ff_trigger

input_absinfo

input_event

input_id

input_keymap_entry

Functions

apm_ioc_standby⚠

apm_ioc_suspend⚠

btrfs_ioc_balance_ctl⚠

btrfs_ioc_clone⚠

btrfs_ioc_default_subvol⚠

btrfs_ioc_get_fslabel⚠

btrfs_ioc_quota_rescan_wait⚠

btrfs_ioc_scrub_cancel⚠

btrfs_ioc_set_fslabel⚠

btrfs_ioc_start_sync⚠

btrfs_ioc_subvol_getflags⚠

btrfs_ioc_subvol_setflags⚠

btrfs_ioc_sync⚠

btrfs_ioc_trans_end⚠

btrfs_ioc_trans_start⚠

btrfs_ioc_wait_sync⚠

cm_iocardoff⚠

cm_iocgstatus⚠

cm_iocsrdr⚠

cxl_ioctl_get_process_element⚠

cz_boot_data⚠

cz_boot_end⚠

cz_boot_start⚠

cz_nboards⚠

cz_test⚠

drm_ioctl_agp_acquire⚠

drm_ioctl_agp_release⚠

drm_ioctl_drop_master⚠

drm_ioctl_mode_rmfb⚠

drm_ioctl_radeon_cp_idle⚠

drm_ioctl_radeon_cp_reset⚠

drm_ioctl_radeon_cp_resume⚠

drm_ioctl_radeon_cp_start⚠

drm_ioctl_radeon_flip⚠

drm_ioctl_radeon_reset⚠

drm_ioctl_radeon_swap⚠

drm_ioctl_set_master⚠

eviocgabs⚠

eviocgbit⚠

eviocgeffects⚠

eviocgid⚠

eviocgkey⚠

eviocgkeycode⚠

eviocgled⚠

eviocgmtslots⚠

eviocgname⚠

eviocgphys⚠

eviocgprop⚠

eviocgrab⚠

eviocgrep⚠

eviocgsnd⚠

eviocgsw⚠

eviocguniq⚠

eviocgversion⚠

eviocrevoke⚠

eviocrmff⚠

eviocsclockid⚠

eviocsff⚠

eviocskeycode⚠

eviocsrep⚠

ext2_ioc_getflags⚠

ext2_ioc_getversion⚠

ext2_ioc_getversion_new⚠

ext2_ioc_group_extend⚠

ext2_ioc_setflags⚠

ext2_ioc_setversion⚠

ext2_ioc_setversion_new⚠

ext4_ioc_resize_fs⚠

fat_ioctl_get_attributes⚠

fat_ioctl_get_volume_id⚠

fat_ioctl_set_attributes⚠

fioclex⚠

fs_ioc32_getflags⚠

fs_ioc32_getversion⚠

fs_ioc32_setflags⚠

fs_ioc32_setversion⚠

fs_ioc_getflags⚠

fs_ioc_getversion⚠

fs_ioc_setflags⚠

fs_ioc_setversion⚠

fw_cdev_ioc_get_speed⚠

hidiocapplication⚠

hidiocgflag⚠

hidiocgrdescsize⚠

hidiocgversion⚠

hidiocinitreport⚠

hidiocsflag⚠

iiocdbgvar⚠

iiocdrvctl⚠

iiocgetcps⚠

iiocgetdvr⚠

iiocgetmap⚠

iiocgetprf⚠

iiocgetset⚠

iiocnetaif⚠

iiocnetaln⚠

iiocnetanm⚠

iiocnetasl⚠

iiocnetdif⚠

iiocnetdil⚠

iiocnetdln⚠

iiocnetdnm⚠

iiocnetdwrset⚠

iiocnetgcf⚠

iiocnetgnm⚠

iiocnetgpn⚠

iiocnethup⚠

iiocnetlcr⚠

iiocnetscf⚠

iiocsetbrj⚠

iiocsetgst⚠

iiocsetmap⚠

iiocsetprf⚠

iiocsetset⚠

iiocsetver⚠

iiocsigprf⚠

ioc_dir
Extracts the "direction" (read/write/none) from an encoded ioctl command.

ioc_nr
Extracts the ioctl number from an encoded ioctl command.

ioc_size
Extracts the size from an encoded ioctl command.

ioc_type
Extracts the type from an encoded ioctl command.

ioctl_wdm_max_command⚠

ivtv_ioc_passthrough_mode⚠

jsiocgaxes⚠

jsiocgaxmap⚠

jsiocgbuttons⚠

jsiocgversion⚠

jsiocsaxmap⚠

kiocsound⚠

meyeioc_qbuf_capt⚠

meyeioc_stillcapt⚠

meyeioc_stilljcapt⚠

meyeioc_sync⚠

mgsl_iocclrmodcount⚠

mgsl_iocgif⚠

mgsl_iocgstats⚠

mgsl_iocgtxidle⚠

mgsl_iocgxctrl⚠

mgsl_iocgxsync⚠

mgsl_ioclooptxdone⚠

mgsl_iocrxenable⚠

mgsl_iocsif⚠

mgsl_iocstxidle⚠

mgsl_iocsxctrl⚠

mgsl_iocsxsync⚠

mgsl_ioctxabort⚠

mgsl_ioctxenable⚠

mgsl_iocwaitevent⚠

nvme_ioctl_id⚠

osiocgnetaddr⚠

osiocsnetaddr⚠

pciioc_controller⚠

pciioc_mmap_is_io⚠

pciioc_mmap_is_mem⚠

pciioc_write_combine⚠

perf_event_ioc_disable⚠

perf_event_ioc_enable⚠

perf_event_ioc_id⚠

perf_event_ioc_period⚠

perf_event_ioc_refresh⚠

perf_event_ioc_reset⚠

perf_event_ioc_set_filter⚠

perf_event_ioc_set_output⚠

pmu_ioc_can_sleep⚠

pmu_ioc_get_backlight⚠

pmu_ioc_get_model⚠

pmu_ioc_grab_backlight⚠

pmu_ioc_has_adb⚠

pmu_ioc_set_backlight⚠

pmu_ioc_sleep⚠

pppiocattach⚠

pppiocattchan⚠

pppiocbundle⚠

pppiocconnect⚠

pppiocdetach⚠

pppiocdisconn⚠

pppiocgasyncmap⚠

pppiocgchan⚠

pppiocgcompressors⚠

pppiocgdebug⚠

pppiocgflags⚠

pppiocgifname⚠

pppiocgmpflags⚠

pppiocgmru⚠

pppiocgrasyncmap⚠

pppiocgunit⚠

pppiocnewunit⚠

pppiocsasyncmap⚠

pppiocscompressor⚠

pppiocsdebug⚠

pppiocsflags⚠

pppiocsmaxcid⚠

pppiocsmpflags⚠

pppiocsmpmru⚠

pppiocsmpmtu⚠

pppiocsmrru⚠

pppiocsmru⚠

pppiocsrasyncmap⚠

pppiocxferunit⚠

reiserfs_ioc_getflags⚠

reiserfs_ioc_getversion⚠

reiserfs_ioc_setversion⚠

reiserfs_ioc_unpack⚠

resierfs_ioc_setflags⚠

rfkill_ioctl_noinput⚠

scsi_ioctl_benchmark_command⚠

scsi_ioctl_doorlock⚠

scsi_ioctl_doorunlock⚠

scsi_ioctl_get_bus_number⚠

scsi_ioctl_get_idlun⚠

scsi_ioctl_probe_host⚠

scsi_ioctl_send_command⚠

scsi_ioctl_start_unit⚠

scsi_ioctl_stop_unit⚠

scsi_ioctl_sync⚠

scsi_ioctl_tagged_disable⚠

scsi_ioctl_tagged_enable⚠

scsi_ioctl_test_unit_ready⚠

siocadddlci⚠

siocaddmulti⚠

siocaddrt⚠

siocaipxitfcrt⚠

siocaipxprislt⚠

siocatalkdifaddr⚠

siocatmark⚠

siocax25addfwd⚠

siocax25adduid⚠

siocax25bpqaddr⚠

siocax25ctlcon⚠

siocax25delfwd⚠

siocax25deluid⚠

siocax25devctl⚠

siocax25getinfo⚠

siocax25getinfoold⚠

siocax25getparms⚠

siocax25getuid⚠

siocax25nouid⚠

siocax25optrt⚠

siocax25setparms⚠

siocbondchangeactive⚠

siocbondenslave⚠

siocbondinfoquery⚠

siocbondrelease⚠

siocbondsethwaddr⚠

siocbondslaveinfoquery⚠

siocbraddbr⚠

siocbraddif⚠

siocbrdelbr⚠

siocbrdelif⚠

siocdarp⚠

siocdeldlci⚠

siocdelmulti⚠

siocdelrt⚠

siocdevplip⚠

siocdevprivate⚠

siocdifaddr⚠

siocdrarp⚠

siocethtool⚠

siocgarp⚠

siocgbpqethparam⚠

siocgcmfirmware⚠

siocgcmfrequency⚠

siocgcmpids⚠

siocgcmstats⚠

siocgdebserint⚠

siocgdongle⚠

siocgetlinkname⚠

siocghwtstamp⚠

siocgifaddr⚠

siocgifbr⚠

siocgifbrdaddr⚠

siocgifconf⚠

siocgifcount⚠

siocgifdstaddr⚠

siocgifencap⚠

siocgifflags⚠

siocgifhwaddr⚠

siocgifindex⚠

siocgifmap⚠

siocgifmem⚠

siocgifmetric⚠

siocgifmtu⚠

siocgifname⚠

siocgifnetmask⚠

siocgifpflags⚠

siocgifslave⚠

siocgiftxqlen⚠

siocgifvlan⚠

siocgkeepalive⚠

siocgkeepperiod⚠

siocglease⚠

siocgmediabusy⚠

siocgmiiphy⚠

siocgmiireg⚠

siocgmode⚠

siocgoutfill⚠

siocgpgrp⚠

siocgpppcstats⚠

siocgpppstats⚠

siocgpppver⚠

siocgqos⚠

siocgrarp⚠

siocgreceiving⚠

siocgstamp⚠

siocgstampns⚠

siocinq⚠

siocipxcfgdata⚠

siocipxncpconn⚠

siocmkclip⚠

siocnrctlcon⚠

siocnrdecobs⚠

siocnrgetparms⚠

siocnrrtctl⚠

siocnrsetparms⚠

siocoutq⚠

siocoutqnsd⚠

siocparm_mask⚠

siocprotoprivate⚠

siocrsaccept⚠

siocrsclrrt⚠

siocrsgcause⚠

siocrsgfacilities⚠

siocrsgl2call⚠

siocrsl2call⚠

siocrsscause⚠

siocrssl2call⚠

siocrtmsg⚠

siocsarp⚠

siocsbandwidth⚠

siocsbpqethaddr⚠

siocsbpqethopt⚠

siocsbpqethparam⚠

siocscmfrequency⚠

siocscmpids⚠

siocsdebserint⚠

siocsdongle⚠

siocsdtrrts⚠

siocshwtstamp⚠

siocsifaddr⚠

siocsifatmtcp⚠

siocsifbr⚠

siocsifbrdaddr⚠

siocsifdstaddr⚠

siocsifencap⚠

siocsifflags⚠

siocsifhwaddr⚠

siocsifhwbroadcast⚠

siocsiflink⚠

siocsifmap⚠

siocsifmem⚠

siocsifmetric⚠

siocsifmtu⚠

siocsifname⚠

siocsifnetmask⚠

siocsifpflags⚠

siocsifslave⚠

siocsiftxqlen⚠

siocsifvlan⚠

siocskeepalive⚠

siocskeepperiod⚠

siocslease⚠

siocsmediabusy⚠

siocsmiireg⚠

siocsmode⚠

siocsoutfill⚠

siocspgrp⚠

siocsrarp⚠

siocwandev⚠

siocx25callaccptapprv⚠

siocx25gcalluserdata⚠

siocx25gcausediag⚠

siocx25gdtefacilities⚠

siocx25gfacilities⚠

siocx25gsubscrip⚠

siocx25scalluserdata⚠

siocx25scausediag⚠

siocx25scudmatchlen⚠

siocx25sdtefacilities⚠

siocx25sendcallaccpt⚠

siocx25sfacilities⚠

siocx25ssubscrip⚠

sndrv_dm_fm_ioctl_clear_patches⚠

sndrv_dm_fm_ioctl_reset⚠

sndrv_dm_fm_ioctl_set_connection⚠

sndrv_dm_fm_ioctl_set_mode⚠

sndrv_dm_fm_oss_ioctl_play_note⚠

sndrv_dm_fm_oss_ioctl_reset⚠

sndrv_dm_fm_oss_ioctl_set_mode⚠

sndrv_dm_fm_oss_ioctl_set_opl⚠

sndrv_dm_fm_oss_ioctl_set_params⚠

sndrv_dm_fm_oss_ioctl_set_voice⚠

sndrv_emu10k1_ioctl_continue⚠

sndrv_emu10k1_ioctl_dbg_read⚠

sndrv_emu10k1_ioctl_pversion⚠

sndrv_emu10k1_ioctl_single_step⚠

sndrv_emu10k1_ioctl_stop⚠

sndrv_emu10k1_ioctl_tram_setup⚠

sndrv_emu10k1_ioctl_zero_tram_counter⚠

sndrv_firewire_ioctl_lock⚠

sndrv_firewire_ioctl_unlock⚠

sndrv_sb_csp_ioctl_pause⚠

sndrv_sb_csp_ioctl_restart⚠

sndrv_sb_csp_ioctl_stop⚠

sndrv_sb_csp_ioctl_unload_code⚠

sonypi_iocgbat1cap⚠

sonypi_iocgbat1rem⚠

sonypi_iocgbat2cap⚠

sonypi_iocgbat2rem⚠

sonypi_iocgbatflags⚠

sonypi_iocgblue⚠

sonypi_iocgbrt⚠

sonypi_iocgfan⚠

sonypi_iocgtemp⚠

sonypi_iocsblue⚠

sonypi_iocsbrt⚠

sonypi_iocsfan⚠

spi_ioc_rd_bits_per_word⚠

spi_ioc_rd_lsb_first⚠

spi_ioc_rd_max_speed_hz⚠

spi_ioc_rd_mode⚠

spi_ioc_rd_mode32⚠

spi_ioc_wr_bits_per_word⚠

spi_ioc_wr_lsb_first⚠

spi_ioc_wr_max_speed_hz⚠

spi_ioc_wr_mode⚠

spi_ioc_wr_mode32⚠

spiocstype⚠

sys_f_ioctlsocket⚠

tioccbrk⚠

tioccons⚠

tiocexcl⚠

tiocgdev⚠

tiocgetd⚠

tiocgexcl⚠

tiocgicount⚠

tiocglcktrmios⚠

tiocgpgrp⚠

tiocgpkt⚠

tiocgptlck⚠

tiocgptn⚠

tiocgrs485⚠

tiocgserial⚠

tiocgsid⚠

tiocgsoftcar⚠

tiocgwinsz⚠

tiocinq⚠

tiocl_blankedscreen⚠

tiocl_blankscreen⚠

tiocl_getfgconsole⚠

tiocl_getkmsgredirect⚠

tiocl_getmousereporting⚠

tiocl_getshiftstate⚠

tiocl_pastesel⚠

tiocl_scrollconsole⚠

tiocl_selbuttonmask⚠

tiocl_selchar⚠

tiocl_selclear⚠

tiocl_selline⚠

tiocl_selloadlut⚠

tiocl_selmousereport⚠

tiocl_selpointer⚠

tiocl_selword⚠

tiocl_setkmsgredirect⚠

tiocl_setsel⚠

tiocl_setvesablank⚠

tiocl_unblankscreen⚠

tioclinux⚠

tiocm_car⚠

tiocm_cd⚠

tiocm_cts⚠

tiocm_dsr⚠

tiocm_dtr⚠

tiocm_le⚠

tiocm_loop⚠

tiocm_out1⚠

tiocm_out2⚠

tiocm_ri⚠

tiocm_rng⚠

tiocm_rts⚠

tiocm_sr⚠

tiocm_st⚠

tiocmbic⚠

tiocmbis⚠

tiocmget⚠

tiocmiwait⚠

tiocmset⚠

tiocnotty⚠

tiocnxcl⚠

tiocoutq⚠

tiocpkt⚠

tiocpkt_data⚠

tiocpkt_dostop⚠

tiocpkt_flushread⚠

tiocpkt_flushwrite⚠

tiocpkt_ioctl⚠

tiocpkt_nostop⚠

tiocpkt_start⚠

tiocpkt_stop⚠

tiocsbrk⚠

tiocsctty⚠

tiocser_temt⚠

tiocserconfig⚠

tiocsergetlsr⚠

tiocsergetmulti⚠

tiocsergstruct⚠

tiocsergwild⚠

tiocsersetmulti⚠

tiocserswild⚠

tiocsetd⚠

tiocsig⚠

tiocslcktrmios⚠

tiocspgrp⚠

tiocsptlck⚠

tiocsrs485⚠

tiocsserial⚠

tiocssoftcar⚠

tiocsti⚠

tiocswinsz⚠

tiocvhangup⚠

ubi_iocdet⚠

ubi_iocebch⚠

ubi_ioceber⚠

ubi_iocebismap⚠

ubi_iocebunmap⚠

ubi_iocrmvol⚠

ubi_iocvolrmblk⚠

ubi_iocvolup⚠

usbtmc_ioctl_abort_bulk_in⚠

usbtmc_ioctl_abort_bulk_out⚠

usbtmc_ioctl_clear⚠

usbtmc_ioctl_clear_in_halt⚠

usbtmc_ioctl_clear_out_halt⚠

usbtmc_ioctl_indicator_pulse⚠

vidioc_g_input⚠

vidioc_g_output⚠

vidioc_g_priority⚠

vidioc_log_status⚠

vidioc_omap3isp_stat_en⚠

vidioc_overlay⚠

vidioc_reserved⚠

vidioc_s_input⚠

vidioc_s_output⚠

vidioc_s_priority⚠

vidioc_streamoff⚠

vidioc_streamon⚠

wdioc_getbootstatus⚠

wdioc_getpretimeout⚠

wdioc_getstatus⚠

wdioc_gettemp⚠

wdioc_gettimeleft⚠

wdioc_gettimeout⚠

wdioc_keepalive⚠

wdioc_setoptions⚠

wdioc_setpretimeout⚠

wdioc_settimeout⚠

x86_ioc_rdmsr_regs⚠

x86_ioc_wrmsr_regs⚠

xfs_ioc_freeze⚠

xfs_ioc_getversion⚠

xfs_ioc_getxflags⚠

xfs_ioc_goingdown⚠

xfs_ioc_setxflags⚠

xfs_ioc_thaw⚠